Use PIMPL design pattern for the RecorderEventNotifier
